Fixed changing avatar & refactored structure
