+++ Car/src/activity/journey/journey-activity/journey-applicant/JourneyApplicant.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { ScrollView, Text, View } from "react-native";
-import UserService from "../../../../../../api-service/user-service/UserService";
-import User from "../../../../../../models/user/User";
-import AvatarLogo from "../../../../../components/avatar-logo/AvatarLogo";
+import UserService from "../../../../../api-service/user-service/UserService";
+import User from "../../../../../models/user/User";
+import AvatarLogo from "../../../../components/avatar-logo/AvatarLogo";
 import JourneyApplicantStyle from "./JourneyApplicantStyle";
-import Indicator from "../../../../../components/activity-indicator/Indicator";
+import Indicator from "../../../../components/activity-indicator/Indicator";
 
 const JourneyApplicant = (props: any) => {
     const { userId } = props.route.params;
