+++ Car/src/activity/journey/journey-activity/journey-applicant/JourneyApplicantStyle.tsx
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import Font from "../../../../../data/fonts/Font";
+import Font from "../../../../data/fonts/Font";
 
 const JourneyApplicantStyle = StyleSheet.create({
     mainContainer: {
