+++ Car/src/activity/journey/journey-activity/journey-page/JourneyPage.tsx
@@ -1,19 +1,19 @@
 import React, { useEffect, useState } from "react";
 import { FlatList, Platform, Text, TouchableOpacity, View } from "react-native";
-import JourneyService from "../../../../../../api-service/journey-service/JourneyService";
-import Stop from "../../../../../../models/stop/Stop";
-import User from "../../../../../../models/user/User";
-import BottomPopup from "../../../../../components/bottom-popup/BottomPopup";
+import JourneyService from "../../../../../api-service/journey-service/JourneyService";
+import Stop from "../../../../../models/stop/Stop";
+import User from "../../../../../models/user/User";
+import BottomPopup from "../../../../components/bottom-popup/BottomPopup";
 import JourneyPageStyle from "./JourneyPageStyle";
-import Journey from "../../../../../../models/Journey";
+import Journey from "../../../../../models/Journey";
 import { useNavigation } from "@react-navigation/native";
 import { LinearTextGradient } from "react-native-text-gradient";
 import { Divider } from "react-native-elements";
 import Moment from "moment";
 import Ionicons from "react-native-vector-icons/Ionicons";
-import AvatarLogo from "../../../../../components/avatar-logo/AvatarLogo";
-import Indicator from "../../../../../components/activity-indicator/Indicator";
-import StopType from "../../../../../../models/stop/StopType";
+import AvatarLogo from "../../../../components/avatar-logo/AvatarLogo";
+import Indicator from "../../../../components/activity-indicator/Indicator";
+import StopType from "../../../../../models/stop/StopType";
 import { ScrollView } from "react-native-gesture-handler";
 
 const JourneyPage = ({ props }: any) => {
