+++ Car/src/activity/journey/journey-activity/map-address/SearchJourneyMap.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import MapView, { Marker, PROVIDER_GOOGLE } from "react-native-maps";
-import { mapStyle } from "../map-address/SearchJourneyMapStyle";
+import { mapStyle } from "./SearchJourneyMapStyle";
 
 const SearchJourneyMap = (props: any) => {
     return (
@@ -16,7 +16,7 @@ const SearchJourneyMap = (props: any) => {
             customMapStyle={mapStyle}
         >
             <Marker
-                image={require("../../../../../../assets/images/custom-marker.png")}
+                image={require("../../../../../assets/images/custom-marker.png")}
                 coordinate={{
                     latitude: props?.latitude,
                     longitude: props?.longitude
