+++ Car/src/activity/my-profile/my-profile-activity/cars/car-activity/add-cars/AddCars.tsx
@@ -13,18 +13,18 @@ import {
     ImagePickerResponse,
     launchImageLibrary
 } from "react-native-image-picker/src";
-import BrandService from "../../../../../../api-service/brand-service/BrandService";
-import CarService from "../../../../../../api-service/car-service/CarService";
-import ModelService from "../../../../../../api-service/model-service/ModelService";
-import CarBrand from "../../../../../../models/car/CarBrand";
-import CarColor from "../../../../../../models/car/CarColor";
-import CarModel from "../../../../../../models/car/CarModel";
-import AuthContext from "../../../../../components/auth/AuthContext";
-import CarDropDownPickerItem from "../../../../../components/car-drop-down-picker/CarDropDownItem";
-import CarDropDownPicker from "../../../../../components/car-drop-down-picker/CarDropDownPicker";
-import CarTextInput from "../../../../../components/car-text-input/CarTextInput";
+import BrandService from "../../../../../../../api-service/brand-service/BrandService";
+import CarService from "../../../../../../../api-service/car-service/CarService";
+import ModelService from "../../../../../../../api-service/model-service/ModelService";
+import CarBrand from "../../../../../../../models/car/CarBrand";
+import CarColor from "../../../../../../../models/car/CarColor";
+import CarModel from "../../../../../../../models/car/CarModel";
+import AuthContext from "../../../../../../components/auth/AuthContext";
+import CarDropDownPickerItem from "../../../../../../components/car-drop-down-picker/CarDropDownItem";
+import CarDropDownPicker from "../../../../../../components/car-drop-down-picker/CarDropDownPicker";
+import CarTextInput from "../../../../../../components/car-text-input/CarTextInput";
 import AddCarsStyle from "./AddCarsStyle";
-import * as navigation from "../../../../../components/navigation/Navigation";
+import * as navigation from "../../../../../../components/navigation/Navigation";
 
 const AddCars = () => {
     const { user } = useContext(AuthContext);
@@ -290,8 +290,7 @@ const AddCars = () => {
                         style={AddCarsStyle.carButtonSave}
                         onPress={() => {
                             if (validateFields()) {
-                                saveCarHandle();
-                                navigation.goBack();
+                                saveCarHandle().then(() => navigation.goBack());
                             }
                         }}
                     >
