+++ Car/src/activity/my-profile/my-profile-activity/settings/settings-tabs/SettingsTabs.tsx
@@ -4,9 +4,9 @@ import { Text, TouchableOpacity, View } from "react-native";
 import Ionicons from "react-native-vector-icons/Ionicons";
 import HeaderStyle from "../../../../../components/styles/HeaderStyle";
 import * as navigation from "../../../../../components/navigation/Navigation";
-import AppSettings from "../app-settings/AppSettings";
-import NotificationSettings from "../notification-settings/NotificationSettings";
-import ChatSettings from "../chat-settings/ChatSettings";
+import AppSettings from "../settings-activity/app-settings/AppSettings";
+import NotificationSettings from "../settings-activity/notification-settings/NotificationSettings";
+import ChatSettings from "../settings-activity/chat-settings/ChatSettings";
 import Settings from "../Settings";
 import RNRestart from "react-native-restart";
 import AsyncStorage from "@react-native-community/async-storage";
